% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tune_classifier_spark.R
\name{detect_cluster_config}
\alias{detect_cluster_config}
\title{Detect Spark Cluster Configuration}
\usage{
detect_cluster_config(sc, verbose = TRUE)
}
\arguments{
\item{sc}{Spark connection}

\item{verbose}{Print information}
}
\value{
List with cluster configuration details
}
\description{
Detect Spark Cluster Configuration
}
