% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tune_classifier_spark.R
\name{.create_spark_estimator}
\alias{.create_spark_estimator}
\title{Create Spark ML Estimator}
\usage{
.create_spark_estimator(sc, learner, seed = 123)
}
\arguments{
\item{sc}{Spark connection}

\item{learner}{Algorithm name}

\item{seed}{Random seed}
}
\value{
Spark ML estimator
}
\description{
Create Spark ML Estimator
}
\keyword{internal}
