% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tune_classifier_spark.R
\name{.train_spark_model}
\alias{.train_spark_model}
\title{Train Spark ML Model with Parameters}
\usage{
.train_spark_model(data, learner, params, seed = 123)
}
\arguments{
\item{data}{Prepared Spark DataFrame}

\item{learner}{Algorithm name}

\item{params}{Parameter list}

\item{seed}{Random seed}
}
\value{
Trained Spark ML model
}
\description{
Train Spark ML Model with Parameters
}
\keyword{internal}
