% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tune_classifier_spark.R
\name{generate_param_combinations}
\alias{generate_param_combinations}
\title{Generate Random Parameter Combinations from ParamSet}
\usage{
generate_param_combinations(param_space, n)
}
\arguments{
\item{param_space}{paradox ParamSet defining the search space}

\item{n}{Number of random combinations to generate}
}
\value{
List of parameter lists
}
\description{
Generate Random Parameter Combinations from ParamSet
}
