% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tune_classifier_spark.R
\name{print.mlr3_ensemble}
\alias{print.mlr3_ensemble}
\title{Print Method for mlr3_ensemble}
\usage{
\method{print}{mlr3_ensemble}(x, ...)
}
\arguments{
\item{x}{mlr3_ensemble object}

\item{...}{Additional arguments}
}
\description{
Print Method for mlr3_ensemble
}
