% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tune_classifier.R,
%   R/auto_tune_classifier_spark.R
\name{get_default_search_space}
\alias{get_default_search_space}
\title{Get Default Search Spaces for Supported Algorithms}
\usage{
get_default_search_space(algorithm)

get_default_search_space(algorithm)
}
\arguments{
\item{algorithm}{Algorithm name ("ranger" or "xgboost")}
}
\value{
ps() object with default search space

paradox ParamSet with default search space
}
\description{
Helper function to get default parameter search spaces for ranger and xgboost
}
\examples{
\dontrun{
ranger_space <- get_default_search_space("ranger")
xgboost_space <- get_default_search_space("xgboost")
}

}
